backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-bigquery
aisdb
syda
polars
deltalake
langchain-openai
langgraph
pandera[polars]